Mksquashfs: Make inode_bytes, inode_size and total_inode_bytes long long
As "unsigned ints" they could occasionally exceed 2GB, but,
%d was used in printf, resulting in the values being
displayed as negative.

While fixing that, make them long long, as they could
exceed 4GB.

+++ squashfs-tools/read_fs.c
@@ -936,7 +936,7 @@ long long read_filesystem(char *root_name, int fd, struct squashfs_super_block *
 	unsigned int *root_inode_size, unsigned int *inode_dir_start_block,
 	int *file_count, int *sym_count, int *dev_count, int *dir_count,
 	int *fifo_count, int *sock_count, long long *uncompressed_file,
-	unsigned int *uncompressed_inode, unsigned int *uncompressed_directory,
+	long long *uncompressed_inode, unsigned int *uncompressed_directory,
 	unsigned int *inode_dir_inode_number,
 	unsigned int *inode_dir_parent_inode,
 	void (push_directory_entry)(char *, squashfs_inode, int, int),
